Common: Fixed applyElementWiseOperator for multidimensional mappings.
When adding multiple mappings over time *and* frequency, the result may
not be correctly represented by simply calculating the values at the entries
of the original mappings. For example, think of two intersecting mappings
over time and frequency as follows:

  +-----+
  |     |
+---------+
| |     | |
+---------+
  |     |
  +-----+

The result must contain entries at the intersections too, because the
interpolation between values calculated by applying the operator at the
original entries (at the + signs) will not give the correct result.